Milestone 1.0

+++ android/Application/src/main/java/de/tudarmstadt/es/able/LeDeviceListAdapter.java
@@ -11,55 +11,90 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Created by user on 14.02.18.
+ * Adapter for holding devices found through scanning.
+ *
+ * @author A. Poljakow, Puria Izady (puria.izady@stud.tu-darmstadt.de)
+ * @version 1.0
  */
-// Adapter for holding devices found through scanning.
+
 class LeDeviceListAdapter extends BaseAdapter {
     private List<BluetoothDevice> mLeDevices;
     private LayoutInflater mInflator;
 
-    //Constructor
+    /**
+     * Initializes the list and the inflater.
+     * @param inflater
+     */
     public LeDeviceListAdapter(LayoutInflater inflater) {
         super();
         mLeDevices = new ArrayList<>();
-        //mInflator = DeviceScanActivity.this.getLayoutInflater();
         mInflator = inflater;
     }
 
-    //Method to add additional bluetoothItems
+    /**
+     * Adds the item device to the list, if it is not inside the list.
+     * @param device is added to the list
+     */
     public void addDevice(BluetoothDevice device) {
         if(!mLeDevices.contains(device)) {
             mLeDevices.add(device);
         }
     }
 
+    /**
+     * Returns the device on the a specific position of the list.
+     * @param position of the device which will be returned
+     * @return device on position
+     */
     public BluetoothDevice getDevice(int position) {
         return mLeDevices.get(position);
     }
 
+    /**
+     * Clears the list.
+     */
     public void clear() {
         mLeDevices.clear();
     }
 
+    /**
+     * @return size of the list.
+     */
     @Override
     public int getCount() {
         return mLeDevices.size();
     }
 
+    /**
+     * Gets the device with the id i.
+     * @param i id of device.
+     * @return device with id i.
+     */
     @Override
     public Object getItem(int i) {
         return mLeDevices.get(i);
     }
 
+    /**
+     * Converts int to long.
+     * @param i
+     * @return
+     */
     @Override
     public long getItemId(int i) {
-        return i;
+        return (long) i;
     }
 
+    /**
+     * Sets the list View of the GUI inside DeviceScanActivity.
+     * @param i
+     * @param view
+     * @param viewGroup
+     * @return
+     */
     @Override
     public View getView(int i, View view, ViewGroup viewGroup) {
         DeviceScanActivity.ViewHolder viewHolder;
-        // General ListView optimization code.
         if (view == null) {
             view = mInflator.inflate(R.layout.listitem_device, null);
             viewHolder = new DeviceScanActivity.ViewHolder();
